+++ src/Macrm.hs
@@ -5,39 +5,15 @@
 
 module Macrm where
 
-import Control.Conditional (ifM)
-import Control.Monad
-  ( foldM,
-    unless,
-    when,
-  )
+import Control.Monad (when)
 import Data.Char (toUpper)
-import Data.Fixed
-  ( Fixed,
-    HasResolution,
-    Uni,
-    showFixed,
-  )
 import Data.Int (Int32)
-import Data.List.Utils (replace)
+import Data.List (intercalate)
 import Data.Maybe
   ( fromJust,
-    isNothing,
+    listToMaybe,
   )
 import qualified Data.Text as T
-import Data.Time.LocalTime
-  ( TimeOfDay (TimeOfDay),
-    ZonedTime
-      ( zonedTimeToLocalTime
-      ),
-    getZonedTime,
-    localTimeOfDay,
-  )
-import Data.Tuple.Utils
-  ( fst3,
-    snd3,
-    thd3,
-  )
 import Data.Version (showVersion)
 import Foreign.C.String (withCString)
 import GitHash
@@ -62,13 +38,7 @@ import System.Console.CmdArgs
     typ,
     (&=),
   )
-import System.Directory
-  ( doesDirectoryExist,
-    getHomeDirectory,
-    listDirectory,
-    renameDirectory,
-    renameFile,
-  )
+import System.Directory (listDirectory)
 import System.Exit
   ( ExitCode
       ( ExitFailure,
@@ -99,12 +69,7 @@ import System.Posix.Files
     groupReadMode,
     groupWriteMode,
     intersectFileModes,
-    isBlockDevice,
-    isCharacterDevice,
     isDirectory,
-    isNamedPipe,
-    isSocket,
-    isSymbolicLink,
     otherExecuteMode,
     otherReadMode,
     otherWriteMode,
@@ -217,55 +182,6 @@ getOptions =
 absolutize :: FilePath -> IO FilePath
 absolutize path = fromJust . guess_dotdot <$> absolute_path path
 
-moveToTrash :: [FileInfo] -> IO ()
-moveToTrash [] = return ()
-moveToTrash (fileInfo : fis) = do
-  pair <- makePair fileInfo
-  move pair
-  moveToTrash fis
-  where
-    makePair :: FileInfo -> IO (FilePath, FilePath)
-    makePair (path, _, _) = do
-      let fileOrDirName = T.unpack . last . T.splitOn "/" . T.pack $ path
-      removedPath <- getRemovedPath fileOrDirName
-      return (path, removedPath)
-    move :: (FilePath, FilePath) -> IO ()
-    move (from, to) =
-      ifM
-        (doesDirectoryExist from)
-        (renameDirectory from to)
-        (renameFile from to)
-
-getRemovedPath :: FilePath -> IO FilePath
-getRemovedPath fileOrDirName = do
-  dayOfTime <- getCurrentDayOfTime
-  searchRemovedPath fileOrDirName $ ' ' : dayOfTime
-
-searchRemovedPath :: FilePath -> String -> IO FilePath
-searchRemovedPath fileOrDirName suffix = do
-  homePath <- getHomeDirectory
-  let trashPath = homePath ++ "/.Trash/"
-  removed <- listDirectory trashPath
-  if fileOrDirName `elem` removed
-    then searchRemovedPath (fileOrDirName ++ suffix) suffix
-    else return $ trashPath ++ fileOrDirName
-
-getCurrentDayOfTime :: IO String
-getCurrentDayOfTime = do
-  zonedTime <- getZonedTime
-  let (TimeOfDay hour minute second) =
-        (localTimeOfDay . zonedTimeToLocalTime) zonedTime
-      sHour = if hour >= 10 then show hour else '0' : show hour
-      sMinute = if minute >= 10 then show minute else '0' : show minute
-      uSecond = changeResolution second :: Uni
-      sSecond' = showFixed True uSecond
-      sSecond = if uSecond >= 10 then sSecond' else '0' : sSecond'
-      suffix = sHour ++ "." ++ sMinute ++ "." ++ sSecond
-  return suffix
-
-changeResolution :: (HasResolution a, HasResolution b) => Fixed a -> Fixed b
-changeResolution = fromRational . toRational
-
 rm :: Options -> ExitCode -> UserID -> [FileInfo] -> [FilePath] -> IO ExitCode
 rm (Options False False False False False False False False []) ExitSuccess _ [] [] =
   do
@@ -281,16 +197,18 @@ rm _ exitCode _ removables [] = do
     _ -> return $ ExitFailure 1
 rm options exitCode uid removables (path : paths) = do
   fileInfo <- getFileInfo path
-  if snd3 fileInfo == NotExists
-    then
+  case fileInfo of
+    (_, NotExists, _) ->
       if force options
         then rm options exitCode uid removables paths
         else do
           hPutStrLn stderr $ "macrm: " ++ path ++ ": No such file or directory"
           rm options (ExitFailure 1) uid removables paths
-    else do
-      let status = fromJust . thd3 $ fileInfo
-      let isDir = snd3 fileInfo == Exists && isDirectory status
+    (_, _, Nothing) -> do
+      hPutStrLn stderr $ "macrm: " ++ path ++ ": unable to inspect file"
+      rm options (ExitFailure 1) uid removables paths
+    (_, fileExists, Just status) -> do
+      let isDir = fileExists == Exists && isDirectory status
       let withRecursive = recursive options || recursive' options
       let withDirectory = directory options
       isNotEmpty <-
@@ -323,14 +241,12 @@ rm options exitCode uid removables (path : paths) = do
             else do
               let fileUid = fileOwner status
                   fileGid = fileGroup status
-              mMessage <-
-                if uid == fileUid
-                  then return Nothing
-                  else Just <$> makeMessage fileInfo fileUid fileGid
               needRemove <-
-                if isNothing mMessage
+                if uid == fileUid
                   then return True
-                  else getAgreement (fromJust mMessage) path
+                  else do
+                    message <- makeMessage path status fileUid fileGid
+                    getAgreement message path
               if needRemove
                 then do
                   when (verbose options) $ putStrLn path
@@ -339,27 +255,15 @@ rm options exitCode uid removables (path : paths) = do
 
 remove :: [FileInfo] -> IO ExitCode
 remove fileInfos = do
-  let (paths, isDeadLinks, mStatuses) =
-        foldl
-          ( \(paths', isDeadLinks', mStatuses') (path, isDeadLink, mStatus) ->
-              (path : paths', isDeadLink : isDeadLinks', mStatus : mStatuses')
-          )
-          ([], [], [])
-          fileInfos
+  let paths = map (\(path, _, _) -> path) fileInfos
   absolutePaths <- mapM absolutize paths
-  (normals, specials) <-
-    foldM filterSpecialFiles ([], []) $
-      zip3 absolutePaths isDeadLinks mStatuses
-  unless (null specials) $ moveToTrash . reverse $ specials
-  if null normals
-    then return ExitSuccess
-    else executeScript . createScript . map fst3 . reverse $ normals
+  executeScript . createScript . reverse $ absolutePaths
 
 executeScript :: String -> IO ExitCode
 executeScript script = do
   (Just stdIn, _, _, ph) <-
     createProcess
-      (proc "osascript" [])
+      (proc "osascript" ["-l", "JavaScript"])
         { std_in = CreatePipe,
           std_out = CreatePipe,
           std_err = CreatePipe
@@ -371,34 +275,39 @@ executeScript script = do
 
 createScript :: [FilePath] -> String
 createScript paths =
-  concat
-    [ "set l to {}\n",
-      concatMap
-        ( \path ->
-            "set end of l to posix file \""
-              ++ replace "\"" "\\\"" path
-              ++ "\" as alias\n"
-        )
-        paths,
-      "tell application \"Finder\"\n",
-      "delete l\n",
-      "end tell\n",
-      "return"
+  unlines
+    [ "ObjC.import('Foundation');",
+      "const fm = $.NSFileManager.defaultManager;",
+      "const paths = [",
+      intercalate ",\n" $ map (("  " ++) . quoteJavaScriptString) paths,
+      "];",
+      "for (const path of paths) {",
+      "  const url = $.NSURL.fileURLWithPath(path);",
+      "  const result = Ref();",
+      "  const err = Ref();",
+      "  const ok = fm.trashItemAtURLResultingItemURLError(url, result, err);",
+      "  if (!ok) {",
+      "    throw ObjC.unwrap(err[0].localizedDescription);",
+      "  }",
+      "}"
     ]
+
+quoteJavaScriptString :: String -> String
+quoteJavaScriptString value = "\"" ++ concatMap escape value ++ "\""
+  where
+    escape :: Char -> String
+    escape '"' = "\\\""
+    escape '\\' = "\\\\"
+    escape '\n' = "\\n"
+    escape '\r' = "\\r"
+    escape '\t' = "\\t"
+    escape c = [c]
 
 getAgreement :: String -> FilePath -> IO Bool
 getAgreement message path = do
   putStr $ message ++ path ++ "? "
   hFlush stdout
-  input <- getLine
-  return $ not (null input) && toUpper (head input) == 'Y'
-
-filterSpecialFiles ::
-  ([FileInfo], [FileInfo]) -> FileInfo -> IO ([FileInfo], [FileInfo])
-filterSpecialFiles (normals, specials) fileInfo =
-  if isSpecialFile fileInfo
-    then return (normals, fileInfo : specials)
-    else return (fileInfo : normals, specials)
+  maybe False ((== 'Y') . toUpper) . listToMaybe <$> getLine
 
 getFileFlags :: FilePath -> IO (Maybe String)
 getFileFlags path = do
@@ -413,12 +322,12 @@ getFileFlags path = do
   output <- hGetContents stdOut
   if null output
     then return Nothing
-    else
-      let flags = T.unpack $ (T.splitOn " " . T.pack $ output) !! 7
-       in return $ Just flags
+    else case drop 4 (words output) of
+      flags : _ -> return $ Just flags
+      [] -> return Nothing
 
-makeMessage :: FileInfo -> UserID -> GroupID -> IO String
-makeMessage (path, _, Just status) uid gid = do
+makeMessage :: FilePath -> FileStatus -> UserID -> GroupID -> IO String
+makeMessage path status uid gid = do
   userAndGroup <- makeUserAndGroupString uid gid
   mFlags <- getFileFlags path
   return $
@@ -428,7 +337,6 @@ makeMessage (path, _, Just status) uid gid = do
       ++ userAndGroup
       ++ maybe "" (" " ++) mFlags
       ++ " for "
-makeMessage _ _ _ = undefined -- never happen
 
 makePermissionString :: FileStatus -> String
 makePermissionString status =
@@ -482,12 +390,12 @@ makeUserAndGroupString uid gid = do
         then name'
         else searchIdName uidOrGid ss
       where
-        splitted :: [T.Text]
-        splitted = T.splitOn ":" . T.pack $ s
         id' :: String
-        id' = T.unpack $ splitted !! 2
         name' :: String
-        name' = T.unpack . head $ splitted
+        (id', name') = case T.splitOn ":" (T.pack s) of
+          nameText : _ : idText : _ ->
+            (T.unpack idText, T.unpack nameText)
+          _ -> ("", "")
 
 getFileInfo :: FilePath -> IO FileInfo
 getFileInfo path = do
@@ -497,17 +405,6 @@ getFileInfo path = do
     else do
       status <- getSymbolicLinkStatus path
       return (path, fileExists, Just status)
-
-isSpecialFile :: FileInfo -> Bool
-isSpecialFile (_, NotExists, _) = False
-isSpecialFile (_, DeadLink, _) = True
-isSpecialFile (_, Exists, Nothing) = undefined -- never happen
-isSpecialFile (_, Exists, Just status) =
-  isSymbolicLink status
-    || isNamedPipe status
-    || isSocket status
-    || isCharacterDevice status
-    || isBlockDevice status
 
 isPathExists :: FilePath -> IO FileExists
 isPathExists path = do
@@ -525,11 +422,11 @@ isPathExists path = do
     close(fd);
     return 2; // exists
   } |]
-  return $ case rc of
-    0 -> NotExists
-    1 -> DeadLink
-    2 -> Exists
-    _ -> undefined -- never happen
+  case rc of
+    0 -> return NotExists
+    1 -> return DeadLink
+    2 -> return Exists
+    _ -> fail $ "unexpected lstat result: " ++ show rc -- never happen
 
 gitInfo :: GitInfo
 gitInfo = $$(tGitInfoCwd)